% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA.r
\name{ICAA}
\alias{ICAA}
\title{The ICA function based on asymmetry}
\usage{
ICAA(X)
}
\value{
S, M and W as in the ICAA paper of Spurek et al. (2017)
}
\description{
The ICA function based on asymmetry
}
\examples{
ICA(data)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Independent}
